requirements.txt
numpy
GitPython